Update ve delete işlemlerine yetkilendirme eklendi...

+++ mern-memories-app-server/routers/memoryRouter.js
@@ -55,6 +55,9 @@ router.put("/:id", auth, async (req, res) => {
     if (!mongoose.Types.ObjectId.isValid(id))
       res.status(404).json({ message: "Memory id is not valid..." });
 
+    const oldMemory = await Memory.findById(id);
+    if (req.creatorId !== oldMemory.creatorId) return res.sendStatus(403);
+
     const { title, content, creator, image } = req.body;
     const updateMemory = await Memory.findByIdAndUpdate(
       id,
@@ -74,6 +77,9 @@ router.delete("/:id", auth, async (req, res) => {
     const { id } = req.params;
     if (!mongoose.Types.ObjectId.isValid(id))
       res.status(404).json({ message: "Memory id is not valid..." });
+
+    const oldMemory = await Memory.findById(id);
+    if (req.creatorId !== oldMemory.creatorId) return res.sendStatus(403);
 
     await Memory.findByIdAndDelete(id);
     res.status(200).json({ message: "Memory has been deleted" });
